+++ src/test/view/reviewCommentController.test.ts
@@ -118,6 +118,7 @@ describe('ReviewCommentController', function () {
 			comments: [],
 			collapsibleState: vscode.CommentThreadCollapsibleState.Expanded,
 			label: 'Start discussion',
+			readOnly: false,
 			dispose: () => { }
 		};
 	}