+++ src/view/fileTypeDecorationProvider.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode';
 import { fromFileChangeNodeUri, fromPRUri } from '../common/uri';
 import { GitChangeType } from '../common/file';
 
-export class FileTypeDecorationProvider implements vscode.DecorationProvider {
+export class FileTypeDecorationProvider implements vscode.FileDecorationProvider {
 	private _disposables: vscode.Disposable[];
 
 	constructor(
@@ -16,15 +16,14 @@ export class FileTypeDecorationProvider implements vscode.DecorationProvider {
 		this._disposables.push(vscode.window.registerDecorationProvider(this));
 	}
 
-	_onDidChangeDecorations: vscode.EventEmitter<vscode.Uri | vscode.Uri[]> = new vscode.EventEmitter<vscode.Uri | vscode.Uri[]>();
-	onDidChangeDecorations: vscode.Event<vscode.Uri | vscode.Uri[]> = this._onDidChangeDecorations.event;
-	provideDecoration(uri: vscode.Uri, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Decoration> {
+	_onDidChange: vscode.EventEmitter<vscode.Uri | vscode.Uri[]> = new vscode.EventEmitter<vscode.Uri | vscode.Uri[]>();
+	onDidChange: vscode.Event<vscode.Uri | vscode.Uri[]> = this._onDidChange.event;
+	provideFileDecoration(uri: vscode.Uri, token: vscode.CancellationToken): vscode.ProviderResult<vscode.FileDecoration> {
 		const fileChangeUriParams = fromFileChangeNodeUri(uri);
 		if (fileChangeUriParams && fileChangeUriParams.status !== undefined) {
 			return {
-				bubble: false,
-				letter: this.letter(fileChangeUriParams.status),
-				priority: 1
+				propagate: false,
+				badge: this.letter(fileChangeUriParams.status)
 			};
 		}
 
@@ -32,9 +31,8 @@ export class FileTypeDecorationProvider implements vscode.DecorationProvider {
 
 		if (prParams && prParams.status !== undefined) {
 			return {
-				bubble: false,
-				letter: this.letter(prParams.status),
-				priority: 1
+				propagate: false,
+				badge: this.letter(prParams.status)
 			};
 		}