+++ src/view/treeDecorationProvider.ts
@@ -6,7 +6,7 @@
 import * as vscode from 'vscode';
 import { fromFileChangeNodeUri } from '../common/uri';
 
-class TreeDecorationProvider implements vscode.DecorationProvider {
+class TreeDecorationProvider implements vscode.FileDecorationProvider {
 	private fileHasComments: Map<string, boolean> = new Map<string, boolean>();
 
 	updateFileComments(resourceUri: vscode.Uri, prNumber: number, fileName: string, hasComments: boolean): void {
@@ -14,22 +14,21 @@ class TreeDecorationProvider implements vscode.DecorationProvider {
 		const oldValue = this.fileHasComments.get(key);
 		if (oldValue !== hasComments) {
 			this.fileHasComments.set(`${prNumber}:${fileName}`, hasComments);
-			this._onDidChangeDecorations.fire(resourceUri);
+			this._onDidChange.fire(resourceUri);
 		}
 	}
 
-	_onDidChangeDecorations: vscode.EventEmitter<vscode.Uri | vscode.Uri[]> = new vscode.EventEmitter<vscode.Uri | vscode.Uri[]>();
-	onDidChangeDecorations: vscode.Event<vscode.Uri | vscode.Uri[]> = this._onDidChangeDecorations.event;
-	provideDecoration(uri: vscode.Uri, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Decoration> {
+	_onDidChange: vscode.EventEmitter<vscode.Uri | vscode.Uri[]> = new vscode.EventEmitter<vscode.Uri | vscode.Uri[]>();
+	onDidChange: vscode.Event<vscode.Uri | vscode.Uri[]> = this._onDidChange.event;
+	provideFileDecoration(uri: vscode.Uri, token: vscode.CancellationToken): vscode.ProviderResult<vscode.FileDecoration> {
 		const query = fromFileChangeNodeUri(uri);
 		if (query) {
 			const key = `${query.prNumber}:${query.fileName}`;
 			if (this.fileHasComments.get(key)) {
 				return {
-					bubble: false,
-					title: 'Commented',
-					letter: '◆',
-					priority: 2
+					propagate: false,
+					tooltip: 'Commented',
+					badge: '◆'
 				};
 			}
 		}